Create "generateNumber" function, delete unnecessary files
In the last commit we had files like index.html, so it was necessary to
delete it due to making the result appear only in the browser.

Already the file "multiple.ts" has been replaced by index.ts, the same
as after the command "tsc" will create an "index.js" with its structure.

The index.js / index.ts file contains multiple number search and replace
functionality.